recsys/botify/requirements.txt
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-And-Redis==1.0.0
python-json-logger==2.0.2
mmh3==3.0.0
numpy
--extra-index-url https://download.pytorch.org/whl/cu116
torch
pytorch-lightning